fix(babel): resolve circular imports
